# docs/swagger.yaml
basePath: /api
definitions:
  entity.EmployeeTaskKanbanEnum:
    enum:
    - TO_DO
    - IN_PROGRESS
    - NEED_REVIEW
    - COMPLETED
    type: string
    x-enum-varnames:
    - EMPLOYEE_TASK_KANBAN_ENUM_TODO
    - EPMLOYEE_TASK_KANBAN_ENUM_IN_PROGRESS
    - EMPLOYEE_TASK_KANBAN_ENUM_NEED_REVIEW
    - EMPLOYEE_TASK_KANBAN_ENUM_COMPLETED
  entity.EmployeeTaskPriorityEnum:
    enum:
    - LOW
    - MEDIUM
    - HIGH
    type: string
    x-enum-varnames:
    - EMPLOYEE_TASK_PRIORITY_ENUM_LOW
    - EMPLOYEE_TASK_PRIORITY_ENUM_MEDIUM
    - EMPLOYEE_TASK_PRIORITY_ENUM_HIGH
  entity.EmployeeTaskStatusEnum:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - EMPLOYEE_TASK_STATUS_ENUM_ACTIVE
    - EMPLOYEE_TASK_STATUS_ENUM_INACTIVE
  entity.TemplateTaskPriorityEnum:
    enum:
    - LOW
    - MEDIUM
    - HIGH
    type: string
    x-enum-varnames:
    - TEMPLATE_TASK_PRIORITY_ENUM_LOW
    - TEMPLATE_TASK_PRIORITY_ENUM_MEDIUM
    - TEMPLATE_TASK_PRIORITY_ENUM_HIGH
  entity.TemplateTaskStatusEnum:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - TEMPLATE_TASK_STATUS_ENUM_ACTIVE
    - TEMPLATE_TASK_STATUS_ENUM_INACTIVE
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        type: integer
    type: object
  request.CreateEmployeeTaskRequest:
    properties:
      cover_path:
        type: string
      description:
        type: string
      employee_id:
        type: string
      employee_task_attachments:
        items:
          $ref: '#/definitions/request.EmployeeTaskAttachmentRequest'
        type: array
      employee_task_checklists:
        items:
          $ref: '#/definitions/request.EmployeeTaskChecklistRequest'
        type: array
      end_date:
        type: string
      name:
        type: string
      priority:
        type: string
      start_date:
        type: string
      template_task_id:
        type: string
    required:
    - cover_path
    - employee_id
    - end_date
    - name
    - priority
    - start_date
    type: object
...
  request.EmployeeTaskAttachmentRequest:
    properties:
      file:
        $ref: '#/definitions/multipart.FileHeader'
      path:
        type: string
    required:
    - file
    type: object
  request.EmployeeTaskChecklistRequest:
    properties:
      id:
        type: string
      is_checked:
        type: string
      name:
        type: string
      verified_by:
        type: string
    required:
    - name
    type: object
...
  request.UpdateEmployeeTaskRequest:
    properties:
      cover_path:
        type: string
      description:
        type: string
      employee_id:
        type: string
      employee_task_attachments:
        items:
          $ref: '#/definitions/request.EmployeeTaskAttachmentRequest'
        type: array
      employee_task_checklists:
        items:
          $ref: '#/definitions/request.EmployeeTaskChecklistRequest'
        type: array
      end_date:
        type: string
      id:
        type: string
      is_done:
        enum:
        - "YES"
        - "NO"
        type: string
      kanban:
        type: string
      name:
        type: string
      notes:
        type: string
      priority:
        type: string
      proof:
        $ref: '#/definitions/multipart.FileHeader'
      proof_path:
        type: string
      start_date:
        type: string
      status:
        type: string
      template_task_id:
        type: string
      verified_by:
        type: string
    required:
    - cover_path
    - employee_id
    - end_date
    - id
    - name
    - priority
    - start_date
    type: object
...
  response.EmployeeTaskAttachmentResponse:
    properties:
      created_at:
        type: string
      employee_task_id:
        type: string
      id:
        type: string
      path:
        type: string
      path_origin:
        type: string
      updated_at:
        type: string
    type: object
  response.EmployeeTaskChecklistResponse:
    properties:
      created_at:
        type: string
      employee_task_id:
        type: string
      id:
        type: string
      is_checked:
        type: string
      name:
        type: string
      updated_at:
        type: string
      verified_by:
        type: string
      verified_by_name:
        type: string
    type: object
  response.EmployeeTaskResponse:
    properties:
      cover_path:
        type: string
      created_at:
        type: string
      description:
        type: string
      employee_id:
        type: string
      employee_name:
        type: string
      employee_task_attachments:
        items:
          $ref: '#/definitions/response.EmployeeTaskAttachmentResponse'
        type: array
      employee_task_checklists:
        items:
          $ref: '#/definitions/response.EmployeeTaskChecklistResponse'
        type: array
      end_date:
        type: string
      id:
        type: string
      is_done:
        type: string
      kanban:
        $ref: '#/definitions/entity.EmployeeTaskKanbanEnum'
      name:
        type: string
      notes:
        type: string
      priority:
        $ref: '#/definitions/entity.EmployeeTaskPriorityEnum'
      proof:
        type: string
      source:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/entity.EmployeeTaskStatusEnum'
      template_task:
        $ref: '#/definitions/response.TemplateTaskResponse'
      template_task_id:
        type: string
      updated_at:
        type: string
      verified_by:
        type: string
      verified_by_name:
        type: string
    type: object
  response.TemplateTaskAttachmentResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      path:
        type: string
      path_origin:
        type: string
      template_task_id:
        type: string
      updated_at:
        type: string
    type: object
  response.TemplateTaskChecklistResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      template_task_id:
        type: string
      updated_at:
        type: string
    type: object
  response.TemplateTaskResponse:
    properties:
      cover_path_origin:
        type: string
      coverPath:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_duration:
        type: integer
      id:
        type: string
      name:
        type: string
      priority:
        $ref: '#/definitions/entity.TemplateTaskPriorityEnum'
      source:
        type: string
      status:
        $ref: '#/definitions/entity.TemplateTaskStatusEnum'
      template_task_attachments:
        items:
          $ref: '#/definitions/response.TemplateTaskAttachmentResponse'
        type: array
      template_task_checklists:
        items:
          $ref: '#/definitions/response.TemplateTaskChecklistResponse'
        type: array
      updated_at:
        type: string
    type: object
...
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Julong Onboarding API Docs
  version: "1.0"
paths:
...
  /employee-tasks:
    get:
      consumes:
      - application/json
      description: Find all employee task paginated
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: Created At
        in: query
        name: created_at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.EmployeeTaskResponse'
      security:
      - BearerAuth: []
      summary: Find all employee task paginated
      tags:
      - Employee Task
    post:
      consumes:
      - multipart/form-data
      description: Create new employee task
      parameters:
      - description: Create Template Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateEmployeeTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.EmployeeTaskResponse'
      security:
      - BearerAuth: []
      summary: Create new employee task
      tags:
      - Employee Task
  /employee-tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete employee task
      parameters:
      - description: Employee Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success delete employee task
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete employee task
      tags:
      - Employee Task
    get:
      consumes:
      - application/json
      description: Find employee task by id
      parameters:
      - description: Employee Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.EmployeeTaskResponse'
      security:
      - BearerAuth: []
      summary: Find employee task by id
      tags:
      - Employee Task
    put:
      consumes:
      - multipart/form-data
      description: Update employee task
      parameters:
      - description: Employee Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Employee Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateEmployeeTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.EmployeeTaskResponse'
      security:
      - BearerAuth: []
      summary: Update employee task
      tags:
      - Employee Task
  /employee-tasks/count:
    get:
      consumes:
      - application/json
      description: Count employee task by kanban and employee id
      parameters:
      - description: Kanban
        in: query
        name: kanban
        required: true
        type: string
      - description: Employee ID
        in: query
        name: employee_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.EmployeeTaskResponse'
      security:
      - BearerAuth: []
      summary: Count employee task by kanban and employee id
      tags:
      - Employee Task
  /employee-tasks/employee:
    get:
      consumes:
      - application/json
      description: Find all employee task by employee id
      parameters:
      - description: Employee ID
        in: query
        name: employee_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.EmployeeTaskResponse'
      security:
      - BearerAuth: []
      summary: Find all employee task by employee id
      tags:
      - Employee Task
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"